Refactor and fix issue when clicking to select value
