+++ packages/slider/src/RangeSlider.tsx
@@ -1,9 +1,10 @@
-import * as React from 'react';
-import { useLayoutEffect } from '@finn-no/fabric-react-utils';
-import { slider as classes } from '@finn-no/fabric-component-classes';
+import React from 'react';
 import { classNames } from '@chbphone55/classnames';
+import { slider as classes } from '@finn-no/fabric-component-classes';
+import { useLayoutEffect } from '@finn-no/fabric-react-utils';
 import { animated, interpolate, useSpring } from 'react-spring';
 import { useDrag } from 'react-use-gesture';
+import { RangeSliderProps } from './props';
 import useInnerWidth from './useInnerWidth';
 import {
     bigStep,
@@ -11,49 +12,8 @@ import {
     nextValue,
     prevValue,
     ratioToValue,
-    Scale,
     valueToRatio,
 } from './utils';
-
-export type RangeSliderProps = {
-    /** String value that labels the slider. */
-    'aria-label'?: [string, string];
-
-    /** Identifies the element that labels the slider. */
-    'aria-labelledby'?: [string, string];
-
-    /** Human readable text alternative for the value. */
-    'aria-valuetext'?: [string, string];
-
-    /** Additional CSS class for the container. */
-    className?: string;
-
-    /** Whether the slider is disabled. */
-    disabled?: boolean;
-
-    /** Handler that is called every time the value of the slider changes. */
-    onInput?: (value: [number, number]) => void;
-
-    /** Handler that is called when the value of the slider has settled. */
-    onChange?: (value: [number, number]) => void;
-
-    /** The greatest value in the range of permitted values. */
-    max?: number;
-
-    /** The lowest value in the range of permitted values. */
-    min?: number;
-
-    /** A d3-scale object for non linear slider scales.
-     * @see d3-scale repository https://github.com/d3/d3-scale
-     */
-    scale?: Scale;
-
-    /** Specifies the value granularity. */
-    step?: number;
-
-    /** The current value */
-    value: [number, number];
-};
 
 enum Handle {
     Lower = 0,
@@ -105,8 +65,10 @@ const RangeSlider = ({
             case 'PageDown':
                 newValue = prevValue(
                     oldValue,
-                    event.key === 'PageDown' ? bigStep(oldValue, step, min, max, scale) : step,
-                    scale
+                    event.key === 'PageDown'
+                        ? bigStep(oldValue, step, min, max, scale)
+                        : step,
+                    scale,
                 );
                 break;
 
@@ -115,8 +77,10 @@ const RangeSlider = ({
             case 'PageUp':
                 newValue = nextValue(
                     oldValue,
-                    event.key === 'PageUp' ? bigStep(oldValue, step, min, max, scale) : step,
-                    scale
+                    event.key === 'PageUp'
+                        ? bigStep(oldValue, step, min, max, scale)
+                        : step,
+                    scale,
                 );
                 break;
             case 'Home':
@@ -270,7 +234,10 @@ const RangeSlider = ({
             />
             <animated.div
                 aria-disabled={disabled}
-                aria-label={props['aria-label']?.[Handle.Lower] ?? (props['aria-labelledby'] ? undefined : 'Fra')}
+                aria-label={
+                    props['aria-label']?.[Handle.Lower] ??
+                    (props['aria-labelledby'] ? undefined : 'Fra')
+                }
                 aria-labelledby={props['aria-labelledby']?.[Handle.Lower]}
                 // the lower handle is limited by the upper handle
                 aria-valuemax={values[Handle.Upper]}
@@ -309,7 +276,10 @@ const RangeSlider = ({
             </animated.div>
             <animated.div
                 aria-disabled={disabled}
-                aria-label={props['aria-label']?.[Handle.Upper] ?? (props['aria-labelledby'] ? undefined : 'Til')}
+                aria-label={
+                    props['aria-label']?.[Handle.Upper] ??
+                    (props['aria-labelledby'] ? undefined : 'Til')
+                }
                 aria-labelledby={props['aria-labelledby']?.[Handle.Upper]}
                 aria-valuemax={max}
                 // the upper handle is limited by the lower handle
