+++ packages/slider/stories/Slider.stories.tsx
@@ -8,6 +8,7 @@ export default metadata;
 export const Regular = () => {
     const [value, setValue] = React.useState(0);
     const [output, setOutput] = React.useState(value);
+
     return (
         <div>
             <output>{output}</output>
@@ -47,6 +48,7 @@ export const RegularScaled = () => {
     ]);
     const [value, setValue] = React.useState(250);
     const [output, setOutput] = React.useState(value);
+
     return (
         <div>
             <output>
@@ -74,7 +76,7 @@ export const Ranged = () => {
     return (
         <div>
             <output aria-live="off">
-                { /* Setting aria-live="off" here since it's a ranged slider. Users of screen readers would want to hear the value of the slider they are currently adjusting, not the value of the entire range. */ }
+                {/* Setting aria-live="off" here since it's a ranged slider. Users of screen readers would want to hear the value of the slider they are currently adjusting, not the value of the entire range. */}
                 {output[0]} - {output[1]} kr
             </output>
             <RangeSlider
